최소 1개의 Scene이 존재하도록 하기

+++ charmander-api/charmander-api-core/src/main/java/org/cm/api/scene/SceneService.java
@@ -34,10 +34,14 @@ public class SceneService {
         return sceneRepository.save(scene);
     }
 
+    // TODO: 동시성 이슈
     public void deleteScene(AuthInfo authInfo, Long projectId, Long sceneId) {
         var scene = sceneRepository.findProjectSceneForUpdate(projectId, sceneId, authInfo.getMemberId())
             .orElseThrow(() -> new CoreApiException(CoreApiExceptionCode.SCENE_NOT_FOUND));
-
+        if (sceneRepository.countByProjectId(scene.getProject().getId()) <= 1) {
+            throw new CoreApiException(CoreApiExceptionCode.SCENE_DELETE_LAST_SCENE);
+        }
+        scene.getProject().removeScene(scene);
         sceneRepository.delete(scene);
     }
